// src/redux/tag-Types.ts
export enum tagTypes {
  specialties = "specialties",
  admin = "admin",
  doctor = "doctor",
  patient = "patient",
  schedule = "schedule",
  appointment = "appointment",
  doctorSchedule = "doctorSchedule",
  user = "user",
  prescription = "prescription",
  review = "review",
  payment = "payment",
}
export const tagTypeList = [
  tagTypes.specialties,
  tagTypes.admin,
  tagTypes.doctor,
  tagTypes.patient,
  tagTypes.schedule,
  tagTypes.appointment,
  tagTypes.doctorSchedule,
  tagTypes.user,
  tagTypes.prescription,
  tagTypes.review,
  tagTypes.payment,
];